+++ data/src/main/java/com/example/data/di/ApiModule.kt
@@ -1,6 +1,6 @@
 package com.example.data.di
 
-import com.example.data.remote.network.signup.SignUpApiService
+import com.example.data.remote.network.SignUpApiService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
